# ee/ingester/pkg/client/api/openapi.yaml
openapi: 3.0.3
info:
  title: Formance Simple ingester Service API
  version: 0.1.0
servers:
- url: http://localhost:8080
paths:
...
  /connectors/{connectorID}:
    delete:
      operationId: deleteConnector
      parameters:
      - description: The connector id
        explode: false
        in: path
        name: connectorID
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Connector deleted
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: General error
      summary: Delete connector
      tags:
      - Connectors
    get:
      operationId: getConnectorState
      parameters:
      - description: The connector id
        explode: false
        in: path
        name: connectorID
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/createConnector_201_response'
          description: Connector information
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: General error
      summary: Get connector state
      tags:
      - Connectors
...
  /pipelines/{pipelineID}:
    delete:
      operationId: deletePipeline
      parameters:
      - description: The pipeline id
        explode: false
        in: path
        name: pipelineID
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Pipeline deleted
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: General error
      summary: Delete pipeline
      tags:
      - Pipelines
    get:
      operationId: getPipelineState
      parameters:
      - description: The pipeline id
        explode: false
        in: path
        name: pipelineID
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/createPipeline_201_response'
          description: Pipeline information
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: General error
      summary: Get pipeline state
      tags:
      - Pipelines
...
  /pipelines/{pipelineID}/resume:
    post:
      operationId: resumePipeline
      parameters:
      - description: The pipeline id
        explode: false
        in: path
        name: pipelineID
        required: true
        schema:
          type: string
        style: simple
      responses:
        "202":
          description: Pipeline resumed
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: General error
      summary: Resume pipeline
      tags:
      - Pipelines
  /pipelines/{pipelineID}/start:
    post:
      operationId: startPipeline
      parameters:
      - description: The pipeline id
        explode: false
        in: path
        name: pipelineID
        required: true
        schema:
          type: string
        style: simple
      responses:
        "202":
          description: Pipeline started
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: General error
      summary: Start pipeline
      tags:
      - Pipelines
...
components:
...
  schemas:
...
    PipelineConfiguration:
      properties:
        module:
          type: string
        connectorID:
          type: string
      required:
      - connectorID
      - module
    ConnectorConfiguration:
      example:
        driver: driver
        config:
          key: ""
      properties:
        driver:
          type: string
        config:
          additionalProperties: true
          type: object
      required:
      - config
      - driver
    Connector:
      allOf:
      - $ref: '#/components/schemas/ConnectorConfiguration'
      - $ref: '#/components/schemas/Connector_allOf'
    Pipeline:
      allOf:
      - $ref: '#/components/schemas/PipelineConfiguration'
      - $ref: '#/components/schemas/Pipeline_allOf'
    State:
      properties:
        label:
          enum:
          - STOP
          - PAUSE
          - INIT
          - READY
          type: string
        cursor:
          type: string
        previousState:
          $ref: '#/components/schemas/State'
      required:
      - label
      type: object
    Error:
      properties:
        errorCode:
          type: string
        errorMessage:
          type: string
      required:
      - errorCode
      - errorMessage
      type: object
...
    createConnector_201_response:
      example:
        data: null
      properties:
        data:
          $ref: '#/components/schemas/Connector'
      required:
      - data
...
    createPipeline_201_response:
      example:
        data: null
      properties:
        data:
          $ref: '#/components/schemas/Pipeline'
      required:
      - data
    Connector_allOf:
      properties:
        id:
          type: string
        createdAt:
          format: date-time
          type: string
      required:
      - createdAt
      - id
      type: object
      example: null
    Pipeline_allOf:
      properties:
        id:
          type: string
        state:
          $ref: '#/components/schemas/State'
        createdAt:
          format: date-time
          type: string
      required:
      - createdAt
      - id
      - state
      type: object
      example: null